fix: only celebrate on full month differences
The month calculation was reversed.

Month difference was being calculated the wrong way round, so giving
answers like -157 instead of 157. Moreover, because the check to
display for full months included a "< 6", aimed at always showing
months in the early days of a celebration, the months were *always*
being shown, even when not a full month.
Never noticed because only really checked in the first few months, and
there are no unit tests for the celebration subsystem, it just prints
the answer at the moment.